Split up test, each needs its own VM-level init
