+++ test/freenet/node/simulator/RealNodeRequestInsertShortTest.java
@@ -1,7 +1,5 @@
 package freenet.node.simulator;
 
-import static org.junit.Assert.assertEquals;
-
 import java.io.IOException;
 
 import org.junit.Test;
@@ -19,23 +17,13 @@ import freenet.support.LoggerHook.InvalidThresholdException;
 import freenet.support.compress.InvalidCompressionCodecException;
 import freenet.support.TestProperty;
 
-public class RealNodeRequestInsertTest {
+public class RealNodeRequestInsertShortTest {
 
     @Test
     public void testSmallNetwork() throws CHKEncodeException, SSKEncodeException, FSParseException, PeerParseException, InvalidThresholdException, NodeInitException, ReferenceSignatureVerificationException, InterruptedException, SimulatorOverloadedException, InvalidCompressionCodecException, IOException, KeyDecodeException, ExitException {
         String[] args = 
                 new String[] {"size=25","degree=5","htl=4","drop=0",
                 "seed=12345","bypass=FAST_QUEUE_BYPASS"};
-        RealNodeRequestInsertTester.run(args);
-    }
-
-    @Test
-    public void testBigNetwork() throws CHKEncodeException, SSKEncodeException, FSParseException, PeerParseException, InvalidThresholdException, NodeInitException, ReferenceSignatureVerificationException, InterruptedException, SimulatorOverloadedException, InvalidCompressionCodecException, IOException, KeyDecodeException, ExitException {
-        if(!TestProperty.EXTENSIVE) return;
-        String[] args = 
-                new String[] {"size=500","degree=10","htl=5","drop=0",
-                "seed=123456","bypass=FAST_QUEUE_BYPASS"};
-        RealNodeRequestInsertTester.TARGET_SUCCESSES = 100;
         RealNodeRequestInsertTester.run(args);
     }
